+++ core/src/lib.rs
@@ -9,11 +9,17 @@ pub mod pb;
 /// Helpers to deal with block sources.
 pub mod block_view;
 
-// Instruction trait to be implemented by all instructions. The trait enables you to wokr on
+// Instruction trait to be implemented by all instructions. The trait enables you to work on
 // a generic instruction type instead of working with either [CompiledInstruction] or [InnerInstruction]
 // model.
 pub trait Instruction {
     fn program_id_index(&self) -> u32;
+
+    // Returns the indices of the accounts that are specified for this instruction. Those are
+    // not the resolved addresses but the indices of the accounts in the transaction message.
+    //
+    // If you come from `all_instructions` method, iterator element given when iterating an
+    // instruction has a `resolved_accounts` method that returns the resolved addresses.
     fn accounts(&self) -> &Vec<u8>;
     fn data(&self) -> &Vec<u8>;
     fn stack_height(&self) -> Option<u32>;
@@ -110,6 +116,18 @@ impl Instruction for &InnerInstruction {
 }
 
 impl ConfirmedTransaction {
+    /// Returns the transaction id as a base58 string. Use [Self::hash] method to get the
+    /// transaction's hash as a byte array if it's what you are after
+    pub fn id(&self) -> String {
+        bs58::encode(self.hash()).into_string()
+    }
+
+    /// Returns the transaction's hash as a byte array. Use [Self::id] method to get the
+    /// transaction's id as a base58 string if it's what you are after.
+    pub fn hash(&self) -> &[u8] {
+        &self.transaction.as_ref().unwrap().signatures[0]
+    }
+
     pub fn resolved_accounts(&self) -> Vec<&Vec<u8>> {
         let meta = self.meta.as_ref().unwrap();
         let mut accounts = vec![];